rename app directory for clarity
