--- shortcake/__init__.py
@@ -1,2 +1,0 @@
-import shortcake.models
-import shortcake.schemas
